+++ ts/packages/core/src/models/Toolkits.ts
@@ -19,7 +19,7 @@ import { ComposioAuthConfigNotFoundError } from '../errors/AuthConfigErrors';
 import { ConnectedAccounts } from './ConnectedAccounts';
 import { ConnectionRequest } from './ConnectionRequest';
 import { telemetry } from '../telemetry/Telemetry';
-import { AuthSchemeEnum, AuthSchemeType } from '../types/authConfigs.types';
+import { AuthConfigListResponse, AuthSchemeEnum, AuthSchemeType } from '../types/authConfigs.types';
 import logger from '../utils/logger';
 /**
  * Toolkits class
@@ -387,20 +387,40 @@ export class Toolkits {
     if (!authConfigIdToUse) {
       // create authConfig using composioManagedAuthSchemes
       if (toolkit.authConfigDetails && toolkit.authConfigDetails.length > 0) {
-        const authConfig = await composioAuthConfig.create(toolkitSlug, {
-          type: 'use_composio_managed_auth',
-          name: `${toolkit.name} Auth Config`,
-        });
-        authConfigIdToUse = authConfig.id;
+        try {
+          const authConfig = await composioAuthConfig.create(toolkitSlug, {
+            type: 'use_composio_managed_auth',
+            name: `${toolkit.name} Auth Config`,
+          });
+          authConfigIdToUse = authConfig.id;
+        } catch (error) {
+          if (error instanceof ComposioClient.APIError && error.status === 400) {
+            throw new ComposioAuthConfigNotFoundError(
+              `No Default auth config found for toolkit ${toolkitSlug}`,
+              {
+                meta: {
+                  toolkitSlug,
+                },
+                cause: error,
+                possibleFixes: [
+                  `Please Create an auth config for the toolkit ${toolkitSlug} via the dashboard`,
+                ],
+              }
+            );
+          }
+          throw error;
+        }
       } else {
-        throw new ComposioAuthConfigNotFoundError('No auth config found for toolkit', {
-          meta: {
-            toolkitSlug,
-          },
-        });
+        throw new ComposioAuthConfigNotFoundError(
+          `No auth configs found for toolkit ${toolkitSlug}`,
+          {
+            meta: {
+              toolkitSlug,
+            },
+          }
+        );
       }
     }
-
     // create the auth config
     const composioConnectedAccount = new ConnectedAccounts(this.client);
     return await composioConnectedAccount.initiate(userId, authConfigIdToUse, {